refactor play components into their own tsx files
